Component | TopoJSON: Refector collideAreaLabels
